Added with-screen-switch
