+++ src/components/app/app.jsx
@@ -4,91 +4,12 @@ import PropTypes from 'prop-types';
 
 import {Type} from '../../data';
 
-import * as Action from '../../reducers/reducer';
-import WelcomeScreen from '../welcome-screen/welcome-screen.jsx';
-import GenreQuestionScreen from '../genre-question-screen/genre-question-screen.jsx';
-import ArtistQuestionScreen from '../artist-question-screen/artist-question-screen.jsx';
-import GameOverScreen from '../game-over-screen/game-over-screen.jsx';
-import WinScreen from '../win-screen/win-screen.jsx';
-
-import withActivePlayer from './../../hocs/with-active-player/with-active-player';
-import withUserAnswer from './../../hocs/with-user-answer/with-user-answer';
-import withTransformProps from './../../hocs/with-transform-props/with-transform-props';
-
-const transformPlayerToAnswer = (props) => {
-  const newProps = Object.assign({}, props, {
-    renderAnswer: props.renderPlayer
-  });
-  delete newProps.renderPlayer;
-
-  return newProps;
-};
-
-const ArtistQuestionScreenWrapped = withActivePlayer(
-    ArtistQuestionScreen);
-
-const GenreQuestionScreenWrapped = withActivePlayer(
-    withUserAnswer(
-        withTransformProps(transformPlayerToAnswer)(GenreQuestionScreen)));
-
 class App extends Component {
-  _getScreen(question) {
-
-    if (!question) {
-      const {step, questions} = this.props;
-      if (step > questions.length - 1) {
-        return <WinScreen/>;
-      } else {
-
-        const {
-          maxMistakes,
-          gameTime,
-          onWelcomeScreenClick
-        } = this.props;
-
-
-        return <WelcomeScreen
-          errorCount={maxMistakes}
-          time={gameTime}
-          handleClick={onWelcomeScreenClick}
-        />;
-      }
-    }
-
-    const {
-      mistakes,
-      maxMistakes,
-      onGenreUserAnswer,
-      onArtistUserAnswer,
-      resetGame
-    } = this.props;
-
-    if (mistakes >= maxMistakes) {
-      return <GameOverScreen
-        onRelaunchButtonClick={resetGame}
-      />;
-    }
-
-    switch (question.type) {
-      case `genre`: return <GenreQuestionScreenWrapped
-        answers={question.answers}
-        question={question}
-        onAnswer={(userAnswer) => onGenreUserAnswer(userAnswer, question, mistakes, maxMistakes)}
-      />;
-
-      case `artist`: return <ArtistQuestionScreenWrapped
-        question={question}
-        onAnswer={(userAnswer) => onArtistUserAnswer(userAnswer, question, mistakes, maxMistakes)}
-      />;
-    }
-
-    return null;
-  }
-
   render() {
     const {
       questions,
       step,
+      renderScreen
     } = this.props;
 
     return (
@@ -122,46 +43,23 @@ class App extends Component {
           </div>
         </header>
 
-        {this._getScreen(questions[step])}
+        {renderScreen(questions[step])}
 
       </section>);
   }
 }
 
 App.propTypes = {
-  mistakes: PropTypes.number.isRequired,
-  maxMistakes: PropTypes.number.isRequired,
-  gameTime: PropTypes.number.isRequired,
   questions: PropTypes.array.isRequired,
   step: PropTypes.number.isRequired,
-  onWelcomeScreenClick: PropTypes.func.isRequired,
-  onGenreUserAnswer: PropTypes.func.isRequired,
-  onArtistUserAnswer: PropTypes.func.isRequired,
-  resetGame: PropTypes.func.isRequired
+  renderScreen: PropTypes.func.isRequired
 };
 
 const mapStateToProps = (state, ownProps) => Object.assign(
-    {}, ownProps, {step: state.step, mistakes: state.mistakes,
-    });
-
-const mapDispatchToProps = (dispatch) => ({
-  onWelcomeScreenClick: () => dispatch(Action.ActionCreator.incrementStep()),
-
-  onGenreUserAnswer: (userAnswer, question, mistakes, maxMistakes) => {
-    dispatch(Action.ActionCreator.incrementStep());
-    dispatch(Action.onGenreUserAnswer(userAnswer, question, mistakes, maxMistakes));
-  },
-
-  onArtistUserAnswer: (userAnswer, question, mistakes, maxMistakes) => {
-    dispatch(Action.ActionCreator.incrementStep());
-    dispatch(Action.onArtistUserAnswer(userAnswer, question, mistakes, maxMistakes));
-  },
-
-  resetGame: () => dispatch(Action.ActionCreator.resetState())
-});
+    {}, ownProps, {step: state.step});
 
 export {App};
 
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default connect(mapStateToProps)(App);
